+++ src/App.js
@@ -7,7 +7,7 @@ import AddNewPlayer from './containers/TeamControl/AddPlayer/AddPlayer';
 import SidePanel from './components/Navigation/SidePanel/SidePanel';
 import Modal from './components/UI/Modal/Modal';
 import "./components/Navigation/SidePanel/SidePanel.css";
-import Aux from './hoc/Aux/Aux';
+import Aux from './hoc/_Aux/_Aux';
 
 class App extends Component {
   state = {
